Ensure that file operations always work correctly without interfering with startup process
When a log file is initialized, there is a window (10 frames)  where properties can be set.
After this period, the log file is locked into the set overwrite mode. Before this period concludes
the user may access features that depend on knowing the state of the log file. This ensures that
the state is consistent during setup.